# apps/api/swagger.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/chat/messages": {
      "post": {
        "operationId": "ChatController_sendMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendChatMessageRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendChatMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/messages/with-image": {
      "post": {
        "operationId": "ChatController_sendMessageWithImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "venueId": {
                    "type": "string"
                  },
                  "userMessage": {
                    "type": "string"
                  },
                  "conversationId": {
                    "type": "string"
                  }
                },
                "required": [
                  "image",
                  "venueId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendChatMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/stream": {
      "post": {
        "operationId": "ChatController_streamMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamChatMessageRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/conversations": {
      "get": {
        "operationId": "ChatController_listConversations",
        "parameters": [
          {
            "name": "venueId",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListConversationItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/conversations/{id}": {
      "get": {
        "operationId": "ChatController_getConversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            }
          },
          {
            "name": "venueId",
            "required": true,
            "in": "query",
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "chat"
        ]
      },
      "delete": {
        "operationId": "ChatController_deleteConversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            }
          },
          {
            "name": "venueId",
            "required": true,
            "in": "query",
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    },
...
  },
  "info": {
    "title": "GM AI API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SendChatMessageRequestDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "userMessage": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8000
          },
          "conversationId": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          }
        },
        "required": [
          "venueId",
          "userMessage"
        ]
      },
      "SendChatMessageResponseDto": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string"
          },
          "assistantMessage": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "content": {
                "type": "string"
              },
              "followUps": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "id",
              "content",
              "followUps"
            ],
            "x-nestjs_zod-self-required": true
          },
          "toolCallLog": {
            "type": "array",
            "items": {}
          },
          "retrievedItemIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "conversationId",
          "assistantMessage",
          "toolCallLog",
          "retrievedItemIds"
        ]
      },
      "StreamChatMessageRequestDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "userMessage": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8000
          },
          "conversationId": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          }
        },
        "required": [
          "venueId",
          "userMessage"
        ]
      },
      "ListConversationItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "x-nestjs_zod-has-null": true
          },
          "venueId": {
            "type": "string",
            "x-nestjs_zod-has-null": true
          },
          "venueName": {
            "type": "string",
            "x-nestjs_zod-has-null": true
          },
          "lastMessageAt": {
            "type": "string",
            "x-nestjs_zod-has-null": true
          },
          "preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-has-null": true,
            "x-nestjs_zod-empty-type": true
          }
        },
        "required": [
          "id",
          "venueId",
          "venueName",
          "lastMessageAt",
          "preview"
        ]
      },
      "ConversationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "x-nestjs_zod-has-null": true
          },
          "venueId": {
            "type": "string",
            "x-nestjs_zod-has-null": true
          },
          "channel": {
            "type": "string",
            "x-nestjs_zod-has-null": true
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "user",
                    "assistant"
                  ]
                },
                "content": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                },
                "retrievedItemIds": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "followUps": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "reasoning": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "parts": {},
                "toolCallLog": {
                  "type": "array",
                  "items": {}
                },
                "feedbackKind": {
                  "anyOf": [
                    {
                      "type": "string",
                      "enum": [
                        "up",
                        "down",
                        "regenerate"
                      ]
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "id",
                "role",
                "content",
                "createdAt",
                "retrievedItemIds"
              ]
            },
            "x-nestjs_zod-has-null": true
          }
        },
        "required": [
          "id",
          "venueId",
          "channel",
          "messages"
        ]
      },
...
    }
  }
}